--- elementary-portfolio-list-app/src/models/portfolio.ts
@@ -1,9 +1,0 @@
-export interface Portfolio {
-  id: number;
-  title: string;
-  author: string;
-  created_at: string;
-  abstruct: string;
-  url: string;
-  source: string;
-}
