+++ elementary-portfolio-list-app/src/models/portfolio2.ts
